+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## [Unreleased]
 
-## [0.1.1] - 2025-07-01
+## [0.1.2] - 2025-07-01
 
-- Fix edit command to work without manual env option [\#1](https://github.com/sixis/rails-conflicted-credentials/pull/1) ([SixiS](https://github.com/sixis))
+* Fix edit command to work with saved internal conflicts [\#2](https://github.com/sixis/rails-conflicted-credentials/pull/2) ([SixiS](https://github.com/sixis))
+* Fix edit command to work without manual env option [\#1](https://github.com/sixis/rails-conflicted-credentials/pull/1) ([SixiS](https://github.com/sixis))
 
 ## [0.1.0] - 2025-06-28
 
-- Initial release copied out of rails fork
+* Initial release copied out of rails fork